+++ src/lib/Avatar/index.spec.js
@@ -14,11 +14,17 @@ describe('tests for <Avatar />', () => {
 
     expect(container.find('.cui-avatar__img').hasClass('cui-avatar__img--hidden')).toEqual(true);
     expect(container.find('.cui-avatar__letter').length).toEqual(1);
-    
+
     container.setState({ isImageLoaded: true });
-    
+
     expect(container.find('.cui-avatar__img').hasClass('cui-avatar__img--hidden')).toEqual(false);
     expect(container.find('.cui-avatar__letter').exists()).toEqual(false);
+  });
+
+  it('should handle isDecrypting prop', () => {
+    const container = mount(<Avatar isDecrypting icon={<Icon name='handset_24' />} />);
+
+    expect(container.find('.cui-decrypting').length).toEqual(1);
   });
 
   it('should handle isOverview prop', () => {
@@ -102,17 +108,17 @@ describe('tests for <Avatar />', () => {
 
       it('when the size is default(medium)', () => {
         const container = mount(<Avatar src="test.png" title="Test Group" type="self" />);
-        
+
         expect(container.find('.cui-icon').props().width).toEqual(18);
         expect(container.find('.cui-avatar__self').length).toEqual(1);
       });
-      
+
       it('when size is 40', () => {
         const container = mount(<Avatar src="test.png" size={40} title="Test Group" type='self'/>);
 
         expect(container.find('.cui-icon').props().width).toEqual(18);
         expect(container.find('.cui-avatar__self').length).toEqual(1);
-      });      
+      });
 
     });
 
@@ -194,7 +200,16 @@ describe('tests for <Avatar />', () => {
       container = mount(<Avatar src="test.png" size={84} title="Test Group" />);
       expect(container.find('.cui-avatar--84').length).toEqual(1);
     });
+  });
 
+  it('should apply clickable class when onClick prop is defined', () => {
+    let container = mount(<Avatar src="test.png" title="Test Group" onClick={()=>{}} />);
+    expect(container.find('.cui-avatar--clickable').length).toEqual(1);
+  });
+
+  it('should not apply clickable class when onClick prop is undefined', () => {
+    let container = mount(<Avatar src="test.png" title="Test Group" />);
+    expect(container.find('.cui-avatar--clickable').length).toEqual(0);
   });
 
   it('should display icon', () => {